+++ src/point.rs
@@ -1,15 +1,15 @@
 use std::ops::{Add, Sub};
 
-use crate::{line::Line, plane::Plane};
+use crate::{line::Line, plane::Plane, Float};
 
 #[derive(Debug, Clone, Copy, PartialEq)]
 #[cfg_attr(feature = "serde", derive(serde::Serialize, serde::Deserialize))]
 #[repr(C)]
 pub struct Point {
-    pub x: f64,
-    pub y: f64,
-    pub z: f64,
-    pub w: f64,
+    pub x: Float,
+    pub y: Float,
+    pub z: Float,
+    pub w: Float,
 }
 impl Default for Point {
     fn default() -> Self {
@@ -60,16 +60,16 @@ impl Point {
         w: 1.0,
     };
 
-    pub const fn new(x: f64, y: f64, z: f64, w: f64) -> Self {
+    pub const fn new(x: Float, y: Float, z: Float, w: Float) -> Self {
         Self { x, y, z, w }
     }
-    pub const fn new_position(x: f64, y: f64, z: f64) -> Self {
+    pub const fn new_position(x: Float, y: Float, z: Float) -> Self {
         Self { x, y, z, w: 1.0 }
     }
-    pub const fn new_direction(x: f64, y: f64, z: f64) -> Self {
+    pub const fn new_direction(x: Float, y: Float, z: Float) -> Self {
         Self { x, y, z, w: 0.0 }
     }
-    pub const fn from_val(v: f64, w: f64) -> Self {
+    pub const fn from_val(v: Float, w: Float) -> Self {
         Self {
             x: v,
             y: v,
@@ -78,7 +78,7 @@ impl Point {
         }
     }
 
-    pub fn into_buffer(&self) -> [f64; 4] {
+    pub fn into_buffer(&self) -> [Float; 4] {
         [self.x, self.y, self.z, self.w]
     }
     pub fn into_buffer_32(&self) -> [f32; 4] {
@@ -137,17 +137,17 @@ impl Point {
         Self::new(self.x.round(), self.y.round(), self.z.round(), self.w)
     }
 
-    pub fn dot(self, rhs: Point) -> f64 {
+    pub fn dot(self, rhs: Point) -> Float {
         self.x * rhs.x + self.y * rhs.y + self.z * rhs.z + self.w * rhs.w
     }
-    pub fn antidot(self, rhs: Point) -> f64 {
+    pub fn antidot(self, rhs: Point) -> Float {
         self.w * rhs.w
     }
 
-    pub fn magnitude_squared(self) -> f64 {
+    pub fn magnitude_squared(self) -> Float {
         self.dot(self) / self.antidot(self)
     }
-    pub fn magnitude(self) -> f64 {
+    pub fn magnitude(self) -> Float {
         self.magnitude_squared().sqrt()
     }
 
@@ -160,7 +160,7 @@ impl Point {
         }
     }
 
-    pub fn dist(self, other: Point) -> f64 {
+    pub fn dist(self, other: Point) -> Float {
         (other - self).magnitude()
     }
